Rename schema-data.ts to schema-data.server.ts
Ensures that it is excluded from the client-side bundle.
