+++ app/routes/docs.schema.stable.$.ts
@@ -1,7 +1,7 @@
 import type { DataFunctionArgs } from '@remix-run/node'
 import { redirect } from '@remix-run/node'
 
-import { getLatestRelease } from '~/lib/schema-data'
+import { getLatestRelease } from '~/lib/schema-data.server'
 
 export async function loader({ params: { '*': path } }: DataFunctionArgs) {
   return redirect(`/docs/schema/${await getLatestRelease()}/${path}`)
