+++ app/routes/schema.stable.$.ts
@@ -7,7 +7,7 @@
  */
 import { type DataFunctionArgs, redirect } from '@remix-run/node'
 
-import { getLatestRelease } from '~/lib/schema-data'
+import { getLatestRelease } from '~/lib/schema-data.server'
 
 export async function loader({ params: { '*': path } }: DataFunctionArgs) {
   return redirect(
